# backend/src/test/resources/openapi/openapi-snapshot.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://127.0.0.1:8090"}],"paths":{"/api/webhooks/bureau":{"post":{"tags":["bureau-callback-controller"],"operationId":"receive","parameters":[{"name":"X-Bureau-Signature","in":"header","required":false,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"string","format":"byte"}}},"required":true},"responses":{"200":{"description":"The answer was applied, or had already been applied by an earlier copy."},"401":{"description":"The signature was absent or does not match; nothing was written.","content":{"application/problem+json":{}}},"404":{"description":"No application carries the reference in this callback.","content":{"application/problem+json":{}}},"422":{"description":"The callback body could not be read.","content":{"application/problem+json":{}}}}}},"/api/signup":{"post":{"tags":["sign-up-controller"],"operationId":"submit","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignUpRequestBody"}}},"required":true},"responses":{"201":{"description":"The application was created and decided.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OnboardingResponse"}}}},"202":{"description":"The application was created and is waiting on the bureau.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OnboardingResponse"}}}},"200":{"description":"An application for this CPF was already under way; nothing was created.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OnboardingResponse"}}}},"409":{"description":"A customer already exists with this CPF or e-mail address.","content":{"application/problem+json":{}}},"422":{"description":"The submitted values could not be accepted.","content":{"application/problem+json":{}}}}}},"/api/version":{"get":{"tags":["version-controller"],"operationId":"version","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VersionResponse"}}}}}}},"/api/signup/{onboardingId}":{"get":{"tags":["sign-up-controller"],"operationId":"status","parameters":[{"name":"onboardingId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Where the application got to.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OnboardingResponse"}}}},"404":{"description":"No application exists with that identifier.","content":{"application/problem+json":{}}},"422":{"description":"The identifier is not a well-formed one.","content":{"application/problem+json":{}}}}}},"/api/me":{"get":{"tags":["me-controller"],"operationId":"me","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MeResponse"}}}}}}},"/api/account/me":{"get":{"tags":["account-controller"],"operationId":"me_1","responses":{"200":{"description":"The caller's account and what it holds.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountResponse"}}}},"401":{"description":"No valid bearer token was presented.","content":{"application/problem+json":{}}},"404":{"description":"The caller holds no account.","content":{"application/problem+json":{}}}}}}},"components":{"schemas":{"SignUpRequestBody":{"type":"object","properties":{"fullName":{"type":"string"},"cpf":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"birthDate":{"type":"string"},"monthlyIncome":{"type":"string"}}},"OnboardingResponse":{"type":"object","properties":{"onboardingId":{"type":"string"},"status":{"type":"string"},"reasonCategory":{"type":"string"}}},"VersionResponse":{"type":"object","properties":{"version":{"type":"string"}}},"MeResponse":{"type":"object","properties":{"username":{"type":"string"}}},"AccountResponse":{"type":"object","properties":{"branch":{"type":"string"},"number":{"type":"string"},"balance":{"type":"string"},"currency":{"type":"string"}}}}}}